pyproject.toml
[project]
name = "note-taking-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.0"
dependencies = [
    "email-validator>=2.2.0",
    "faker>=37.3.0",
    "google-genai[aiohttp]>=1.25.0",
    "pillow>=11.2.1",
    "python-dotenv>=1.0.1",
    "python-telegram-bot>=22.1",
    "sqlmodel>=0.0.24",
]
